opp_seeker/requirements.txt
beautifulsoup4==4.13.4
fake_useragent==2.0.2
Flask==3.1.0
geopy==2.4.1
jellyfish==1.2.0
jsonpath==0.82.2
jsonpath_ng==1.7.0
jsonpath_python==1.0.6
psycopg2_binary==2.9.10
Requests==2.32.3
scipy==1.15.2
selenium==4.31.0
googletrans==4.0.2
apscheduler==3.11.0
flask
jellyfish
scipy
selenium
Flask-APScheduler
pytz
pandas
